prime<signal.h>(ctrlc)
Write a function checks if the entered integer is prime or not.
The function should return nonzero if the integer is prime, zero if the integer is not prime.
Use the function in a program that receives integers from user and tells if the integer is prime or not, repeatedly.
The program can be terminated only by pressing CTRL+C.